refactor(test): split shared Vitest runtimes
