+++ apps/api/package.json
@@ -29,6 +29,7 @@
     "server-only": "0.0.1"
   },
   "devDependencies": {
+    "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/react": "19.2.17",
     "@vercel/config": "^0.5.5",
