+++ packages/ai/package.json
@@ -30,6 +30,7 @@
     "parse-domain": "^8.3.1"
   },
   "devDependencies": {
+    "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/react": "19.2.17",
     "@types/react-dom": "^19.2.3",
