+++ packages/ai/vitest.config.ts
@@ -1,32 +1,19 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import { defineConfig } from "vitest/config";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
+import config from "@repo/testing/node";
+import { mergeConfig } from "vitest/config";
 
 const defaultExcludes = ["**/node_modules/**", "coverage/**"];
 
-const config = defineConfig({
+export default mergeConfig(config, {
   test: {
     coverage: {
-      enabled: true,
-      provider: "istanbul",
       reportsDirectory: "./coverage/agents",
       thresholds: {
         100: true,
         perFile: true,
       },
     },
-    environment: "node",
     exclude: defaultExcludes,
     include: ["**/*.test.ts"],
     name: "agents",
   },
-  resolve: {
-    alias: {
-      "@repo": path.resolve(__dirname, "../"),
-    },
-  },
 });
-
-export default config;
