+++ packages/analytics/package.json
@@ -17,6 +17,7 @@
     "server-only": "0.0.1"
   },
   "devDependencies": {
+    "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/react": "19.2.17",
     "@types/react-dom": "19.2.3",
