+++ packages/backend/vitest.config.ts
@@ -1,15 +1,12 @@
 import path from "node:path";
-import { fileURLToPath } from "node:url";
-import { defineConfig } from "vitest/config";
+import config from "@repo/testing/node";
+import { mergeConfig } from "vitest/config";
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
 const defaultExcludes = ["**/node_modules/**", "coverage/**"];
 
-const config = defineConfig({
+export default mergeConfig(config, {
   test: {
     coverage: {
-      enabled: true,
-      provider: "istanbul",
       reportsDirectory: "./coverage",
     },
     // Keep CPU available for Convex Edge VMs when Turbo runs package tests together.
@@ -38,10 +35,7 @@ const config = defineConfig({
   },
   resolve: {
     alias: {
-      "@": path.resolve(__dirname, "./"),
-      "@repo": path.resolve(__dirname, "../"),
+      "@": path.resolve(import.meta.dirname, "./"),
     },
   },
 });
-
-export default config;
