+++ packages/contents/package.json
@@ -23,6 +23,7 @@
     "unified": "11.0.5"
   },
   "devDependencies": {
+    "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@types/mdast": "4.0.4",
     "@types/react": "^19.2.17",
