+++ packages/design-system/vitest.config.ts
@@ -1,21 +1,10 @@
 import path from "node:path";
-import { fileURLToPath } from "node:url";
-import { defineConfig } from "vitest/config";
+import config from "@repo/testing/react";
+import { mergeConfig } from "vitest/config";
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-const config = defineConfig({
-  css: {
-    postcss: {
-      /** Vitest stubs CSS imports; unit tests should not load app Tailwind. */
-      plugins: [],
-    },
-  },
+export default mergeConfig(config, {
   test: {
-    environment: "jsdom",
     coverage: {
-      enabled: true,
-      provider: "istanbul",
       thresholds: {
         100: true,
         perFile: true,
@@ -24,10 +13,7 @@ const config = defineConfig({
   },
   resolve: {
     alias: {
-      "@": path.resolve(__dirname, "./"),
-      "@repo": path.resolve(__dirname, "../"),
+      "@": path.resolve(import.meta.dirname, "./"),
     },
   },
 });
-
-export default config;
