+++ packages/math/package.json
@@ -14,6 +14,7 @@
     "effect": "catalog:"
   },
   "devDependencies": {
+    "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@vitest/coverage-istanbul": "^4.1.10",
     "vitest": "^4.1.10"
