+++ packages/math/vitest.config.ts
@@ -1,29 +1,16 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import { defineConfig } from "vitest/config";
+import config from "@repo/testing/node";
+import { mergeConfig } from "vitest/config";
 
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-const config = defineConfig({
+export default mergeConfig(config, {
   test: {
     coverage: {
-      enabled: true,
-      provider: "istanbul",
       reportsDirectory: "./coverage",
       thresholds: {
         100: true,
         perFile: true,
       },
     },
-    environment: "node",
     include: ["**/*.test.ts"],
     name: "math",
   },
-  resolve: {
-    alias: {
-      "@repo": path.resolve(__dirname, "../"),
-    },
-  },
 });
-
-export default config;
