+++ packages/next-config/package.json
@@ -9,6 +9,7 @@
   },
   "devDependencies": {
     "@next/mdx": "16.2.11",
+    "@repo/testing": "workspace:*",
     "@repo/typescript-config": "workspace:*",
     "@vitest/coverage-istanbul": "^4.1.10",
     "next": "16.2.11",
