--- packages/testing/index.ts
@@ -1,31 +1,0 @@
-import path from "node:path";
-import { fileURLToPath } from "node:url";
-import react from "@vitejs/plugin-react";
-import { defineConfig } from "vitest/config";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-const config = defineConfig({
-  css: {
-    postcss: {
-      /** Vitest stubs CSS imports; unit tests should not load app Tailwind. */
-      plugins: [],
-    },
-  },
-  plugins: [react()],
-  test: {
-    environment: "jsdom",
-    coverage: {
-      enabled: true,
-      provider: "istanbul",
-    },
-  },
-  resolve: {
-    alias: {
-      "@": path.resolve(__dirname, "./"),
-      "@repo": path.resolve(__dirname, "../../packages"),
-    },
-  },
-});
-
-export default config;
