+++ packages/testing/node.ts
@@ -1,0 +1,8 @@
+import { mergeConfig } from "vitest/config";
+import config from "#testing/base";
+
+export default mergeConfig(config, {
+  test: {
+    environment: "node",
+  },
+});
